+++ src/main/java/ch/raising/models/InvestorType.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class InvestorType {
-    private int id;
+    private long id;
     private String name;
     private String description;
 
-    public InvestorType(int id, String name, String description) {
+    public InvestorType(long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
